Added lookupAndTraverse() to the storage interface
In a lot of cases, the query starts from an index lookup, then traverse out
from the result nodes. That is especially common in knowledge graph. In the
current storage interface, that has to be two remote calls, which could be
combined into one to save the network traffic and latency.

This PR is to introduce the new method ```lookupAndTraverse()``` just for
that purpose.